+++ src/tickets/ZendeskView.tsx
@@ -1,29 +1,12 @@
 import React from "react";
 import { view } from "@aha-app/react-easy-state";
-import { useEffect } from "react";
 import TicketsTable from "../components/TicketsTable";
-import { loadViewData } from "../store";
-import { DashboardView, View, ViewData } from "../types";
+import { View } from "../types";
 import { SectionTitle } from "./SectionTitle";
+import { useViewData } from "../lib/useViewData";
 
-const ZendeskView = ({
-  dashboardView,
-  data: { loading, data: viewData } = { loading: false, data: null },
-  view,
-  onRemove,
-}: {
-  dashboardView: DashboardView;
-  data: { loading: boolean; data: ViewData | null } | null;
-  view: View;
-  onRemove: () => void;
-}) => {
-  const fetchData = !!viewData && !loading;
-
-  useEffect(() => {
-    if (!fetchData) {
-      loadViewData(dashboardView.id);
-    }
-  }, [fetchData, dashboardView.id]);
+const ZendeskView = ({ view, onRemove }: { view: View; onRemove: () => void }) => {
+  const { loading, viewData } = useViewData(view);
 
   return (
     <aha-panel>
